Entity moved to a separate artefact References: #44
