+++ MIGRATION_5.md
@@ -2,8 +2,9 @@
 
 # Imports
 
-| Artefact                            | Old package                 | New package                  |
-| ----------------------------------- | --------------------------- | ---------------------------- |
-| LceState and extensions             | com.motorro.rxlcemodel.base | com.motorro.rxlcemodel.lce   |
-| LceUseCase, LceModel and extensions | com.motorro.rxlcemodel.base | com.motorro.rxlcemodel.rx    |
-| Clock, Logger                       | com.motorro.rxlcemodel.base | com.motorro.rxlcemodel.utils |
+| Artefact                            | Old package                        | New package                         |
+| ----------------------------------- | ---------------------------------- | ----------------------------------- |
+| LceState and extensions             | com.motorro.rxlcemodel.base        | com.motorro.rxlcemodel.lce          |
+| LceUseCase, LceModel and extensions | com.motorro.rxlcemodel.base        | com.motorro.rxlcemodel.rx           |
+| Clock, Logger                       | com.motorro.rxlcemodel.base        | com.motorro.rxlcemodel.utils        |
+| Entity, EntityValidatorFactory      | com.motorro.rxlcemodel.base.entity | com.motorro.rxlcemodel.cache.entity |
