+++ disklrucache/src/main/kotlin/com/motorro/rxlcemodel/disklrucache/DiskLruCacheSyncDelegate.kt
@@ -14,8 +14,8 @@
 package com.motorro.rxlcemodel.disklrucache
 
 import com.jakewharton.disklrucache.DiskLruCache
+import com.motorro.rxlcemodel.cache.entity.Entity
 import com.motorro.rxlcemodel.disklrucache.DiskLruCacheSyncDelegate.DiskLruCacheProvider
-import com.motorro.rxlcemodel.rx.entity.Entity
 import com.motorro.rxlcemodel.rx.service.CacheDelegateSerializerDeserializer
 import com.motorro.rxlcemodel.rx.service.SyncDelegateCacheService
 import com.motorro.rxlcemodel.utils.Clock
