+++ disklrucache/src/main/kotlin/com/motorro/rxlcemodel/disklrucache/factory.kt
@@ -13,9 +13,9 @@
 
 package com.motorro.rxlcemodel.disklrucache
 
+import com.motorro.rxlcemodel.cache.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.EntityValidatorFactory
 import com.motorro.rxlcemodel.disklrucache.DiskLruCacheSyncDelegate.DiskLruCacheProvider
-import com.motorro.rxlcemodel.rx.entity.Entity
-import com.motorro.rxlcemodel.rx.entity.EntityValidatorFactory
 import com.motorro.rxlcemodel.rx.service.*
 import java.io.Serializable
 import java.security.MessageDigest
