+++ disklrucache/src/test/kotlin/com/motorro/rxlcemodel/disklrucache/DiskLruCacheSyncDelegateTest.kt
@@ -13,8 +13,8 @@
 
 package com.motorro.rxlcemodel.disklrucache
 
-import com.motorro.rxlcemodel.rx.entity.Entity
-import com.motorro.rxlcemodel.rx.entity.EntityValidator
+import com.motorro.rxlcemodel.cache.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.EntityValidator
 import com.motorro.rxlcemodel.rx.service.CacheDelegateSerializerDeserializer
 import com.motorro.rxlcemodel.utils.Clock
 import com.nhaarman.mockitokotlin2.*
