+++ kserializer/src/main/kotlin/com/motorro/rxlcemodel/kserializer/factory.kt
@@ -13,12 +13,12 @@
 
 package com.motorro.rxlcemodel.kserializer
 
+import com.motorro.rxlcemodel.cache.entity.EntityValidatorFactory
 import com.motorro.rxlcemodel.disklrucache.DiskLruCacheSyncDelegate
 import com.motorro.rxlcemodel.disklrucache.DiskLruCacheSyncDelegate.DiskLruCacheProvider
 import com.motorro.rxlcemodel.disklrucache.createDefaultDelegatePrefix
 import com.motorro.rxlcemodel.disklrucache.createDelegate
 import com.motorro.rxlcemodel.disklrucache.createNormalizedDelegate
-import com.motorro.rxlcemodel.rx.entity.EntityValidatorFactory
 import com.motorro.rxlcemodel.rx.service.CacheFriend
 import com.motorro.rxlcemodel.rx.service.SyncDelegateCacheService
 import com.motorro.rxlcemodel.rx.service.stringifyParams
