+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/StrategyUpdateWrapper.kt
@@ -1,6 +1,6 @@
 package com.motorro.rxlcemodel.rx
 
-import com.motorro.rxlcemodel.rx.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.Entity
 import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.utils.Logger
 import io.reactivex.rxjava3.core.Completable
