+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/UpdateWrapper.kt
@@ -13,8 +13,8 @@
 
 package com.motorro.rxlcemodel.rx
 
+import com.motorro.rxlcemodel.cache.entity.Entity
 import com.motorro.rxlcemodel.lce.LceState
-import com.motorro.rxlcemodel.rx.entity.Entity
 import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.rx.service.UpdateOperationState
 import com.motorro.rxlcemodel.rx.service.UpdateOperationState.*
