+++ rx/src/test/kotlin/com/motorro/rxlcemodel/rx/UpdateWrapperTest.kt
@@ -13,9 +13,9 @@
 
 package com.motorro.rxlcemodel.rx
 
+import com.motorro.rxlcemodel.cache.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.EntityValidator
 import com.motorro.rxlcemodel.lce.LceState
-import com.motorro.rxlcemodel.rx.entity.Entity
-import com.motorro.rxlcemodel.rx.entity.EntityValidator
 import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.rx.service.UpdatingNetService
 import com.motorro.rxlcemodel.rx.service.UpdatingServiceSet
