+++ rx/src/test/kotlin/com/motorro/rxlcemodel/rx/UpdatingLceModelWrapperTest.kt
@@ -13,10 +13,10 @@
 
 package com.motorro.rxlcemodel.rx
 
+import com.motorro.rxlcemodel.cache.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.EntityValidator
 import com.motorro.rxlcemodel.lce.LceState.Content
 import com.motorro.rxlcemodel.lce.LceState.Loading
-import com.motorro.rxlcemodel.rx.entity.Entity
-import com.motorro.rxlcemodel.rx.entity.EntityValidator
 import com.nhaarman.mockitokotlin2.verify
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.schedulers.Schedulers
