+++ rx/src/test/kotlin/com/motorro/rxlcemodel/rx/service/CacheFriendDelegateTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Nikolai Kotchetkov.
+ * Copyright 2022 Nikolai Kotchetkov.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,9 +13,9 @@
 
 package com.motorro.rxlcemodel.rx.service
 
-import com.motorro.rxlcemodel.rx.entity.Entity
-import com.motorro.rxlcemodel.rx.entity.EntityValidator
-import com.motorro.rxlcemodel.rx.entity.EntityValidatorFactory
+import com.motorro.rxlcemodel.cache.entity.Entity
+import com.motorro.rxlcemodel.cache.entity.EntityValidator
+import com.motorro.rxlcemodel.cache.entity.EntityValidatorFactory
 import com.nhaarman.mockitokotlin2.*
 import org.junit.Before
 import org.junit.Test
