+++ sample/src/test/kotlin/com/motorro/rxlcemodel/sample/service/usecase/PatchNoteTest.kt
@@ -13,8 +13,8 @@
 
 package com.motorro.rxlcemodel.sample.service.usecase
 
-import com.motorro.rxlcemodel.rx.entity.EntityValidator
-import com.motorro.rxlcemodel.rx.entity.EntityValidatorFactory
+import com.motorro.rxlcemodel.cache.entity.EntityValidator
+import com.motorro.rxlcemodel.cache.entity.EntityValidatorFactory
 import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.sample.NOTE
 import com.motorro.rxlcemodel.sample.domain.data.NoteList
