+++ settings.gradle
@@ -4,3 +4,4 @@ include ':viewmodel'
 include ':composeview'
 include ':lce'
 include ':utils'
+include ':cache'
